pyproject.toml
[project]
name = "final-proj-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "beautifulsoup4==4.13.4",
    "bs4==0.0.2",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "faiss-cpu==1.11.0",
    "fastapi==0.115.12",
    "filelock==3.18.0",
    "fsspec==2025.5.1",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.1",
    "google-api-python-client==2.172.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.73.0",
    "grpcio-status==1.71.0",
    "h11==0.16.0",
    "hf-xet==1.1.3",
    "httplib2==0.22.0",
    "huggingface-hub==0.33.0",
    "idna==3.10",
    "jinja2==3.1.6",
    "joblib==1.5.1",
    "markupsafe==3.0.2",
    "mpmath==1.3.0",
    "networkx==3.5",
    "numpy==2.3.0",
    "packaging==25.0",
    "pillow==11.2.1",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyparsing==3.2.3",
    "python-dotenv>=1.1.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.4",
    "rsa==4.9.1",
    "safetensors==0.5.3",
    "scikit-learn==1.7.0",
    "scipy==1.15.3",
    "sentence-transformers==4.1.0",
    "setuptools==80.9.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "starlette==0.46.2",
    "sympy==1.14.0",
    "threadpoolctl==3.6.0",
    "tokenizers==0.21.1",
    "torch==2.7.1",
    "tqdm==4.67.1",
    "transformers==4.52.4",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "uritemplate==4.2.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.3",
]
